cloud_function/requirements.txt
cachetools==5.5.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.4.0 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and (platform_system == "Windows" or sys_platform == "win32")
contourpy==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
cycler==0.12.1 ; python_version >= "3.12" and python_version < "4.0"
deprecated==1.2.15 ; python_version >= "3.12" and python_version < "4.0"
fonttools==4.55.2 ; python_version >= "3.12" and python_version < "4.0"
google-api-core==2.23.0 ; python_version >= "3.12" and python_version < "4.0"
google-api-core[grpc]==2.23.0 ; python_version >= "3.12" and python_version < "4.0"
google-auth==2.36.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-appengine-logging==1.5.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-audit-log==0.3.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-bigquery==3.27.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-core==2.4.1 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-logging==3.11.3 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-storage==2.19.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud==0.34.0 ; python_version >= "3.12" and python_version < "4.0"
google-crc32c==1.6.0 ; python_version >= "3.12" and python_version < "4.0"
google-resumable-media==2.7.2 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos==1.66.0 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos[grpc]==1.66.0 ; python_version >= "3.12" and python_version < "4.0"
grpc-google-iam-v1==0.13.1 ; python_version >= "3.12" and python_version < "4.0"
grpcio-status==1.68.1 ; python_version >= "3.12" and python_version < "4.0"
grpcio==1.68.1 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
importlib-metadata==8.5.0 ; python_version >= "3.12" and python_version < "4.0"
kiwisolver==1.4.7 ; python_version >= "3.12" and python_version < "4.0"
loguru==0.7.2 ; python_version >= "3.12" and python_version < "4.0"
matplotlib==3.9.3 ; python_version >= "3.12" and python_version < "4.0"
numpy==2.1.3 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-api==1.28.2 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.12" and python_version < "4.0"
pandas==2.2.3 ; python_version >= "3.12" and python_version < "4.0"
pillow==11.0.0 ; python_version >= "3.12" and python_version < "4.0"
polars==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
proto-plus==1.25.0 ; python_version >= "3.12" and python_version < "4.0"
protobuf==5.29.1 ; python_version >= "3.12" and python_version < "4.0"
pyasn1-modules==0.4.1 ; python_version >= "3.12" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.12" and python_version < "4.0"
pyparsing==3.2.0 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
pytz==2024.2 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.12" and python_version < "4"
six==1.17.0 ; python_version >= "3.12" and python_version < "4.0"
tqdm==4.67.1 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2024.2 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.2.3 ; python_version >= "3.12" and python_version < "4.0"
win32-setctime==1.1.0 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "win32"
wrapt==1.17.0 ; python_version >= "3.12" and python_version < "4.0"
zipp==3.21.0 ; python_version >= "3.12" and python_version < "4.0"
